nft-service/requirements.txt
asgiref             3.8.1
boto3               1.38.23
botocore            1.38.23
Django              5.2.1
djangorestframework 3.16.0
dnspython           2.7.0
jmespath            1.0.1
pip                 24.3.1
pymongo             4.13.0
python-dateutil     2.9.0.post0
python-decouple     3.8
s3transfer          0.13.0
six                 1.17.0
sqlparse            0.5.3
urllib3             2.4.0
